// Atspm/WebUI/src/features/charts/purdueSplitFailure/types.ts
// #region license
// Copyright 2026 Utah Departement of Transportation
// for WebUI - types.ts
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//http://www.apache.org/licenses/LICENSE-2.
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
// #endregion
import {
  BaseChartData,
  BaseChartOptions,
  BasePlan,
  ChartType,
  DataPoint,
} from '@/features/charts/common/types'

export interface PurdueSplitFailureChartOptions extends BaseChartOptions {
  firstSecondsOfRed: number
  metricTypeId: number
  getPermissivePhase: boolean
}

export interface PurdueSplitFailureChartOptionsDefaults {
  firstSecondsOfRed: { id: number; value: string; option: string }
  metricTypeId: { id: number; value: string; option: string }
  getPermissivePhase: { id: number; value: boolean; option: string }
}

export interface PurdueSplitFailurePlan extends BasePlan {
  totalCycles: number
  failsInPlan: number
  percentFails: number
}

interface FailLine {
  timestamp: string
}

export interface RawPurdueSplitFailureData extends BaseChartData {
  approachId: number
  approachDescription: string
  phaseNumber: number
  phaseType: string
  totalSplitFails: number
  plans: PurdueSplitFailurePlan[]
  failLines: FailLine[]
  gapOutGreenOccupancies: DataPoint[]
  gapOutRedOccupancies: DataPoint[]
  forceOffGreenOccupancies: DataPoint[]
  forceOffRedOccupancies: DataPoint[]
  averageGor: DataPoint[]
  averageRor: DataPoint[]
  percentFails: DataPoint[]
}

export interface RawPurdueSplitFailureResponse {
  type: ChartType.PurdueSplitFailure
  data: RawPurdueSplitFailureData[]
}